refactor: rename initializer package to strategy
